lib<appname>.so.* --> lib<appname>.so*
To easy packaging while we are using this build system,
we should have two .control files per application, e.g.

* opie-embeddedkonsole-standalone.control and
* opie-embeddedkonsole-quicklaunch.control.

Both control files would 'provide' the application and it is
up to the packager (i.e. the people who do the feeds) to decide which
ones to ship. This would also easy the packaging of feeds while
in the transition phase of converting arbitrary apps from must-be-standalone
to can-be-standalone-or-quicklaunch.

Comments?