This patch makes the famous Documents Tab customizably optional.
Note that the document tab is so tight coupled to the rest
of the launcher that it proved impossible for me to remove the
tab as is. However I could make all the scanning and the displaying
optional.

The customization can be done via the launcher settings applications -
a new doc tab category has been added. Further customization can be
added to that tab.

I also added a doc tab settings application for use with the "first usage wizard"