OColorButton::~OColorButton ( ) { +    delete d->m_menu;      delete d; }
That was a tricky one to find. If a dynamically loaded shared object (dso)
creates QObjects/QWidgets on the destruction of QApplication they will
be freed. For normal applications these dso's have already been removed
from the address-space leading to calling delete to or from a bogus
part of memory leading to segfaults