This script collects all control files locates Files and Package entrie and generates using that information a makefile (on stdout)
SOME DOCUMENTATION (sent to all developers)

That makefile generates a new packages for a particular control file
ONLY IF

  1. the control file changed
  2. one of the files in the control file changed (including
     support for wildcards).

The script is called automatically when needed by the main makefile
when either

  make packages-mt
  make packages
  make package-SomeControlBaseName

is called.

if a goal that does not begin with package... is called (like calling
make clean or just make) then the generation is NOT performed (even if
the special makefile to be generated does not exist) NOR will that
special makefile be included (if it exists).

This guarantees the performance of the build to remain the same

packages built are generated to the Packages subdirectory in $(TOPDIR)

The makefile does not try to package applications that are not compiled

mkipk might fail if files are missing.  However the makefile is written
such that the packaging continues

Old packages of the same control are removed

EXTRAVERSION is updated automatically (and updated ONLY for packages
that need to be rebuilt).

IMPORTANT :

  Because of the dangling link to /opt/QtPalmtop/bin/quicklauncher
  in bin for all quicklauncher apps.  the makefile exists with
  an error message if /opt... does not exist.  to make the makefile
  work  issue

    mkdir /opt
    ln -sf $(OPIEDIR) /opt/QtPalmtop