Do not scale the buttons in the decoration over and over again.
The logic was if the requested size is not the size of the pixmap
load and scale it to the right size. The problem was the scaleButton method
was changed to not scale the pixmap (to avoid bluring al will kill me for that)
if the delta of the sizes was <= 4.
We now do not try to scale if the delta is <= 4