opiebluez: add scanning.
Ok, guys, everything until now was easy. It gets very ugly to go from here - even
to just get the name of a remote device, you have to setup some filters on the bluez socket,
fill in some random flags to generate a PDU that calls 'get name' and then afterwards
poll until the result comes in. Nasty :/

The BlueZ kernel interface seems to be very badly (if at all) documented.
All people are assuming that you use libbluetooth to talk to that stack.
However since libbluetooth is GPL, we can't do that :/
Guess, we are stuck here until someone finds time and/or motivation to look
into that and create some easy-to-understand examples for how to talk
directly to the BlueZ kernel interface.