This commit makes all supported models bit unique from each other within their respective families. It also makes the model families bit unique with each other. This patch makes button (and general device features) not overlap. This fixes Opie bug 1836.
Thanks goes out to Paul E. (aka BlueLightning) for originally reporting
this problem and providing the text for a reasonable near-term solution. On
top of all that, Paul even test the patch that this commit is based on,
confirming that it was a reasonable solution.

For those of you out there who support distros for the other model families,
please be careful. And note that we have not tested on any of model family
but iPAQs. We don't anticipate any problems, but be careful out there.