This commit is based on the patch submitted to bug 1702: http://opie-bugs.oszine.de/view.php?id=1702
The problem behavior said that the day abbreviations in the PIM suite were
wrong for at least it and fr. The fix is to make sure that there is not a
translation of the abbreviations because the abbreviations have already
been translated.

Thanks goes to an anonymous reporter for reporting and to antonus for
submitting a patch. Thanks to Paul E. (aka BlueLightning) for bringing the
bug to our attention.