The const on the return type is useless.
When you put const on the return type of a function and the return
value is passed by value (and not by reference) then the directive
is excessive (what do you mean the caller can't change a value that
was copied) and on some compilers (like GCC) it will just ignore
the directive. Best to just leave it off.