Refactor and revise code in image effects classes
I am trying to unravel the evaluation order of the code that
deals w/ the opacity of images. It is proving difficult. So I am
trying to refactor it to make it more readable. Hoepfully the
evaluation order will be obvious. Otherwise, it will at least
look a little more spiffy.