Remove assignments that are never used.
An unfortunate side effect of a QUI toolkit that only requires you to
declare and allocate elements to see them is that the usual idiom of
assigning the allocation to something is wasteful. It is wasteful
because the assignment takes up resources and the variable that is
assigned takes up resources, but they are never used. Better to
just allocate and move on. Which is exactly what this commit does.
Along w/ removing any other wasteful assignments.