Add new opie bluetooth daemon
Add a daemon to handle bluetooth adapter control/status, utilising the BlueZ
4.x DBus interface. This takes care of detecting an adapter, enabling and
disabling power on it, and notifying other applications via QCop when its
status changes. There is also now a "temporary" enabled mode which allows
for applications that only need to turn bluetooth on temporarily (e.g. to
beam a file) to do so without turning off the adapter afterwards if it was
already on. As a bonus enabling/disabling of the adapter will now be
possible without having the applet running.

Signed-off-by: Paul Eggleton <bluelightning@bluelightning.org>