launcher: add applet-specific notify popup
Add QPE/TaskBar message(QString,QString) QCop call. First argument is the
message, second argument is the name of the applet to show the notification
for (from the library name, object name or class name of the applet), or
"inputmethod" for the input method selector. If the specified applet can't
be found then the notification is shown for the O menu (as with
message(QString)).

Signed-off-by: Paul Eggleton <bluelightning@bluelightning.org>