libopiemm2: add playback synchronisation
I don't expect many will use PulseAudio on an Opie system but in this
case it highlighted the lack of synchronisation between frames written
to the sound device and the ones coming out of the speakers. Without
PulseAudio this lack of synchronisation didn't seem to be too much of a
problem, but with it the playback position indicator zoomed ahead to the
end and then the application froze until after playback is completed.

There are still a couple of issues with playback with PulseAudio but the
most serious one is now fixed.

Signed-off-by: Paul Eggleton <bluelightning@bluelightning.org>