addressbook: improve handling of organization-only contacts
If Full Name == Organization, allow free modification of the
Organisation field instead of only being able to type it out once (the
old code broke if you pressed backspace), and set the last name as the
organisation. Also provide a more appropriate default for the "File As"
field for these type of contacts.

Fixes bug #1846.

Signed-off-by: Paul Eggleton <bluelightning@bluelightning.org>