opie-eye: add easy access to popup menu in fullscreen mode
When in fullscreen mode with a large image, it can be difficult to
get back out of fullscreen mode without hardware buttons, as activating
the popup menu is made harder by the entire surface being scrollable.
Add a button which appears when the screen is tapped which gives easy
access to the popup menu.

Fixes bug #1824.

Signed-off-by: Paul Eggleton <bluelightning@bluelightning.org>