connmanapplet/connmansettings: adapt to ConnMan API changes
Adapt to changes in the ConnMan technology, manager and service APIs in
version 0.79+. In the process, a few nasty synchronous DBus calls were
removed and some code tidied up.

The applet and settings application now share some source files; I opted
to do this rather than having to create a separate library, but it would
be easy to split them out into one in future.

Signed-off-by: Paul Eggleton <bluelightning@bluelightning.org>