libopie2: implement generic backlight code
Implement backlight control generically instead of having slightly
different code to do the same thing in all ODevice implementations. We
simply look for a backlight device in sysfs and use the first one if one
is found. This means the backlight should now work on any device where
the kernel supports the standard sysfs backlight interface.

Signed-off-by: Paul Eggleton <bluelightning@bluelightning.org>