libopie2: implement generic LED support
Replace the device-specific LED code with generic code that looks for
LED devices in /sys/class/leds. This required changing some of the API
as LEDs are now identified by their names from sysfs rather than members
of a fixed enum.

Applications that just wish to show a basic notification should use the
QCop interface (QPE/TaskBar setLed(int,bool) and new setLed(int,int))
rather than calling ODevice::setLedState directly, as the launcher is
now in charge of determining which LED should be used (by reading a
value from qpe.conf with the intention that this should be configurable
somewhere in future; in the absence of that setting we just look for a
LED with "mail" in its name.) The first parameter is 0 for the default
LED if available, or 1+ for a specific LED - previously this parameter
was ignored.

Additionally, LED blinking is now handled within ODevice rather than
relying on support from the hardware or driver, since this does not seem
to be very common). Two additonal modes were added: "Led_BlinkSuperSlow"
for use with buzzers, and "Led_BlinkSuperFast" for fast eye-catching LED
blinking.

Signed-off-by: Paul Eggleton <bluelightning@bluelightning.org>