libopie2: load buttons from default config file
Instead of hardcoding device buttons in the form of constants, read them
from a file. This means buttons can be handled generically and we avoid
having the button definitons for all machines included on every device.

Note: when deploying to a target device, the appropriate file from
etc/buttons should be copied to qpeDir()/etc/opie_buttons.conf.

Signed-off-by: Paul Eggleton <bluelightning@bluelightning.org>